ji-helper: Improve matching algorithm (weighted preferred ratios)
Each chromatic interval has a preferred ratio, and a weight.
If the ratio for that interval found in the scale matches
preferred, the weight is added; otherwise, 0. Higher sum wins.
This resolves a conflict where e.g. C-D-A would have an
arbitrary tuning (but we really want +comma to win).